+++ test/MasterGenesisSetup.sol
@@ -23,7 +23,7 @@ contract MasterGenesisSetup is MasterSetup {
         uint256 quorumThreshold;
         uint8 requiredKeyTag;
         bytes32 validatorsSszMRoot;
-        uint128 verificationType;
+        uint32 verificationType;
         uint8 version;
     }
